backend/requirements.txt
# FinOS Backend Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
cryptography>=42.0.0
python-multipart>=0.0.9
httpx>=0.27.0

# Production (optional)
# asyncpg>=0.29.0        # PostgreSQL async driver
# alembic>=1.13.0        # Database migrations
# gunicorn>=22.0.0       # Production WSGI server
# redis>=5.0.0           # Caching
# celery>=5.3.0          # Background tasks
